+++ app/src/main/AndroidManifest.xml
@@ -21,7 +21,7 @@
             android:name=".View.ProdutoListActivity"
             android:exported="false" />
         <activity
-            android:name=".View.SplashActivity"
+            android:name=".View.MainScreen"
             android:exported="true"
             android:theme="@style/Theme.AfimdeFeiraXTheme">
             <intent-filter>
@@ -31,13 +31,13 @@
             </intent-filter>
         </activity>
 
-        <activity
-            android:name=".View.LoginActivity"
-            android:exported="true"></activity>
-        <activity
-            android:name=".View.MainActivity"
-            android:exported="true"
-            android:theme="@style/Theme.AfimdeFeiraX.NoActionBar"></activity>
+<!--        <activity-->
+<!--            android:name=".View.LoginActivity"-->
+<!--            android:exported="true"></activity>-->
+<!--        <activity-->
+<!--            android:name=".View.MainScreen"-->
+<!--            android:exported="true"-->
+<!--            android:theme="@style/Theme.AfimdeFeiraX.NoActionBar"></activity>-->
 
         <meta-data
             android:name="com.google.android.geo.API_KEY"
