+++ app/src/main/java/com/example/afimdefeirax/View/Components/MenuComponent.kt
@@ -1,0 +1,5 @@
+package com.example.afimdefeirax.View.Components
+
+import androidx.compose.ui.graphics.vector.ImageVector
+
+data class MenuBar( val  name : String, val icons : ImageVector)
